remove SoundFurniture interface
